+++ src/app/code/community/EbayEnterprise/RiskInsight/Model/Person/Name.php
@@ -26,8 +26,9 @@ class EbayEnterprise_RiskInsight_Model_Person_Name
 	/** @var string */
 	protected $_firstName;
 
-	public function __construct()
+	public function __construct(array $initParams=array())
 	{
+		parent::__construct($initParams);
 		$this->_extractionPaths = array(
 			'setFirstName' => 'string(x:FirstName)',
 		);