+++ src/app/code/community/EbayEnterprise/RiskInsight/Test/Model/Process/ResponseTest.php
@@ -56,10 +56,12 @@ class EbayEnterprise_RiskInsight_Test_Model_Process_ResponseTest
 	protected function _loadXmlTestString($fixtureFile)
 	{
 		$dom = new DOMDocument();
-		$dom->preserveWhiteSpace = false;
 		$dom->load($fixtureFile);
-		$string = $dom->C14N();
-		return $string;
+		$dom->encoding = 'utf-8';
+		$dom->formatOutput = false;
+		$dom->preserveWhiteSpace = false;
+		$dom->normalizeDocument();
+		return $dom->saveXML();
 	}
 
 	/**